+++ src/base/velocity.h
@@ -31,7 +31,7 @@ BEGIN_C_DECLS
  * velocity。
  */
 typedef struct _velocity_t {
-  uint32_t time;
+  uint64_t time;
   uint32_t degree;
 
   float x;
@@ -57,13 +57,13 @@ velocity_t* velocity_reset(velocity_t* v);
  * 更新位置。
  *
  * @param {velocity_t*} v velocity对象。
- * @param {uint32_t} time 时间。
+ * @param {uint64_t} time 时间。
  * @param {xy_t} x x坐标。
  * @param {xy_t} y y坐标。
  *
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
  */
-ret_t velocity_update(velocity_t* v, uint32_t time, xy_t x, xy_t y);
+ret_t velocity_update(velocity_t* v, uint64_t time, xy_t x, xy_t y);
 
 END_C_DECLS
